refactor(cli): organize client commands
