+++ app/src/cli/args.rs
@@ -7,7 +7,7 @@ use std::{
 
 use crate::{
   cli::{
-    commands::server,
+    commands::{client, server},
     environment_target,
     environment_target::EnvironmentTarget,
     secret_format::SecretFormat,
@@ -16,6 +16,7 @@ use crate::{
 };
 
 pub use server::{ServerCommand, ServerLaunchArgs, ServerStartArgs};
+pub use client::ClientCommand;
 
 #[derive(Parser, Debug)]
 #[command(
@@ -120,7 +121,7 @@ pub enum Command {
   /// Connect the CLI to a Dopbase server (`client connect <url>`).
   ///
   /// Without a saved server, client commands use http://localhost:8840.
-  #[command(after_help = CLIENT_HELP)]
+  #[command(after_help = client::HELP)]
   Client {
     #[command(subcommand)]
     command: ClientCommand,
@@ -139,7 +140,7 @@ pub enum Command {
   #[command(after_help = LOGOUT_HELP)]
   Logout,
   /// Alias for `dopbase client status`.
-  #[command(after_help = STATUS_HELP)]
+  #[command(after_help = client::STATUS_ALIAS_HELP)]
   Status,
   /// Create a project, its first environment, and import secrets.
   #[command(after_help = INIT_HELP)]
@@ -294,24 +295,6 @@ pub enum Command {
   },
 }
 
-#[derive(Subcommand, Debug)]
-pub enum ClientCommand {
-  /// Validate a server URL and save it as the active server.
-  ///
-  /// Domains must include http:// or https://. Bare IP addresses use HTTP.
-  /// Accepts the `local` alias to return to the implicit local default
-  /// (http://localhost:8840). Changing servers requires interactive
-  /// confirmation, stops the current managed background server, clears the
-  /// saved CLI session and default, and then requires a new login.
-  #[command(after_help = CLIENT_CONNECT_HELP)]
-  Connect {
-    /// Server URL or bare IP to save, or `local` to use http://localhost:8840.
-    server_url: String,
-  },
-  /// Show the active server, connection state, login, and default environment.
-  #[command(after_help = CLIENT_STATUS_HELP)]
-  Status,
-}
 #[derive(Subcommand, Debug)]
 pub enum ProjectCommand {
   /// Create an empty project.
